Move processing out of the `server` package This moves things away from the `definition.go` file where everything is happening The new `processing` package only deals with the jsonnet `ast` package and doesn't import the `protocol`. All conversions happen on the server level now and don't make it down into processing
